Changed corpus construction mechanism to chunks instead of singular files
